[Chore] 컴포넌트 이름, props 변경
